requirements.txt
streamlit
numpy
scikit-learn
joblib
plotly.graph_objects